Publish pointer file
Initially hidden to avoid people from messing up online lobbies. Now published to let people help with development more easily and since this file being hidden or not has 0 influence on people messing with online lobbies. They'll do it anyways, whether or not I publish this one file or not.